chore(utils): Synchronize file name with function name
