+++ packages/common/utils/src/index.ts
@@ -11,7 +11,6 @@ export * from './delay';
 export * from './device/index';
 export * from './difference';
 export * from './escapeHTML';
-export * from './format';
 export * from './generateID';
 export * from './get-set';
 export * from './getScrollDiffFromBottom';
@@ -24,6 +23,7 @@ export * from './identity';
 export * from './isNotNil';
 export * from './mapValues';
 export * from './Masker';
+export * from './maskName';
 export * from './maxBy';
 export * from './minBy';
 export * from './noop';
